requirements.txt
fastapi==0.111.0
pandas==2.0.1
pmdarima==2.0.2
psycopg2==2.9.9
psycopg2_binary==2.9.9
pydantic==1.9.2
PyYAML==6.0.1
PyYAML==5.4.1
PyYAML==6.0.1
Requests==2.32.2
scikit_learn==1.3.2
